Extract `searchFromLine` and `extractResult` from `index.tsx`
