+++ src/components/JSONInspector/JSONSearch/index.tsx
@@ -2,8 +2,8 @@ import cnx from 'classnames/bind'
 import styles from '../JSONInspector.module.css'
 import { type ChangeEvent, useLayoutEffect, useState } from 'react'
 import { useJSONInspector } from '../JSONInspectorContext'
-import type { IndexedJSONLine } from '../JSONRenderer/types'
 import type { JSONSearchResult } from '../types'
+import { extractResult, searchFromLine } from './logic'
 
 const cx = cnx.bind(styles)
 
@@ -150,75 +150,4 @@ export function JSONSearch({}: Props) {
       </button>
     </fieldset>
   )
-}
-
-type SearchFromLineParams = {
-  line: IndexedJSONLine
-  keyword: string
-  isMatchCase: boolean
-  isMatchWord: boolean
-  isRegexUsed: boolean
-}
-
-function searchFromLine({
-  line,
-  keyword,
-  isMatchCase,
-  isMatchWord,
-  isRegexUsed,
-}: SearchFromLineParams) {
-  const regexp = createRegExp(isRegexUsed, keyword, isMatchWord, isMatchCase)
-
-  if (!regexp) return []
-
-  return line.tokens
-    .map((token) => {
-      const target = token.content
-
-      return {
-        token,
-        match: regexp.exec(target)!,
-      }
-    })
-    .filter(({ match }) => match)
-}
-
-function createRegExp(
-  isRegexUsed: boolean,
-  keyword: string,
-  isMatchWord: boolean,
-  isMatchCase: boolean,
-) {
-  const escapedKeyword = isRegexUsed
-    ? keyword
-    : keyword.replaceAll(/([.?!()\[\]*])/g, '\\$1')
-
-  try {
-    const regexp = new RegExp(
-      isMatchWord ? `(?<!\\w)${escapedKeyword}(?!\\w)` : escapedKeyword,
-      !isMatchCase ? 'i' : undefined,
-    )
-
-    return regexp
-  } catch (e) {
-    return undefined
-  }
-}
-
-function extractResult(match: RegExpExecArray) {
-  if (match.length === 1) {
-    return [[match.index, match.index + match[0].length]]
-  }
-
-  const results: [number, number][] = []
-
-  for (let i = 1; i <= match.length - 1; ++i) {
-    const matchedIndices = match.indices?.[i]
-
-    if (matchedIndices) {
-      results.push(matchedIndices)
-    }
-  }
-
-  return results
 }
